requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
dspy-ai
